[chore] 일기 수정 로직 DetailDiary로 이동
