+++ Nabi/presentation/src/main/java/com/nabi/nabi/views/diary/detail/DetailDiaryViewModel.kt
@@ -9,6 +9,7 @@ import com.nabi.domain.repository.DataStoreRepository
 import com.nabi.domain.usecase.bookmark.AddBookmarkUseCase
 import com.nabi.domain.usecase.bookmark.DeleteBookmarkUseCase
 import com.nabi.domain.usecase.diary.GetDiaryDetailUseCase
+import com.nabi.domain.usecase.diary.UpdateDiaryUseCase
 import com.nabi.nabi.utils.UiState
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.launch
@@ -17,6 +18,7 @@ import javax.inject.Inject
 @HiltViewModel
 class DetailDiaryViewModel @Inject constructor(
     private val getDiaryDetailUseCase: GetDiaryDetailUseCase,
+    private val updateDiaryUseCase: UpdateDiaryUseCase,
     private val addBookmarkUseCase: AddBookmarkUseCase,
     private val deleteBookmarkUseCase: DeleteBookmarkUseCase,
     private val dataStoreRepository: DataStoreRepository
@@ -34,6 +36,8 @@ class DetailDiaryViewModel @Inject constructor(
     private val _deleteState = MutableLiveData<UiState<Unit>>(UiState.Loading)
     val deleteState: LiveData<UiState<Unit>> get() = _deleteState
 
+    private val _updateState = MutableLiveData<UiState<Unit>>(UiState.Loading)
+    val updateState: LiveData<UiState<Unit>> get() = _updateState
 
     fun fetchData(diaryId: Int) {
         _diaryState.value = UiState.Loading
@@ -47,6 +51,21 @@ class DetailDiaryViewModel @Inject constructor(
                     _diaryState.value = UiState.Success(it)
                 }.onFailure { e ->
                     _diaryState.value = UiState.Failure(message = e.message.toString())
+                }
+        }
+    }
+
+    fun updateDiary(id: Int, content: String) {
+        _updateState.value = UiState.Loading
+
+        viewModelScope.launch {
+            val accessToken = dataStoreRepository.getAccessToken().getOrNull().orEmpty()
+
+            updateDiaryUseCase(accessToken, id, content)
+                .onSuccess {
+                    _addState.value = UiState.Success(Unit)
+                }.onFailure { e ->
+                    _addState.value = UiState.Failure(message = e.message.toString())
                 }
         }
     }
